Customer-Service-Implementation/src/main/java/com/example/customerservice/service/CustomerService.java
package com.example.customerservice.service;

import java.util.List;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import com.example.customerservice.entity.Customer;
import com.example.customerservice.repository.CustomerRepository;

@Service
public class CustomerService {

	@Autowired
	private CustomerRepository customerRepository;

	public Customer addCustomer(Customer customer) {
		return customerRepository.save(customer);
	}

	public Customer getCustomerById(Long customerId) {
		return customerRepository.findById(customerId).orElseThrow(() -> new RuntimeException("Customer not found"));
	}

	public List<Customer> getAllCustomers() {
		return customerRepository.findAll();
	}

	public Customer updateCustomer(Long customerId, Customer updatedCustomer) {
		Customer customer = getCustomerById(customerId);
		customer.setName(updatedCustomer.getName());
		customer.setEmail(updatedCustomer.getEmail());
		customer.setPhone(updatedCustomer.getPhone());
		return customerRepository.save(customer);
	}

	public void deleteCustomer(Long customerId) {
		customerRepository.deleteById(customerId);
	}
}
